Added JWT layer
Added the jwt filter, which will authenticate based on the cookie that the browser provides. If the user logs in/registers, then a cookie containing jwt will be created and passed to the browser.
P.S. The JwtCookie filter is out of use. I will decide what to do with it later on